CreateNewFontInLibrary checks for duplicate font names.
If CreateNewFontInLibrary is asked to create a font in a library
in which there is already a font with the specified name, it will
append the index number of the new font to the end of the name.
This ensures that all font names in the libraries are unique. This
should only ever be a problem if something else has been done wrong
somewhere.